pyproject.toml
[project]
name = "self-healing-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "chromadb>=1.5.8",
    "langchain>=1.2.16",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langgraph>=1.1.10",
    "python-dotenv>=1.2.2",
]
